feat: MOve facebook button to bots view
